devide ui test
